Implement TagPredicateRegistry, new superclass for RegistryPredicateRegistry
Seems its trivial to do tags for DamageSources but hard to do set matches. given we don't *need* set matches, not a huge problem
